+++ ov_core/src/cpi/CpiV1.h
@@ -64,7 +64,7 @@ namespace ov_core {
          * @param sigma_ab accelerometer random walk (m/s^3/sqrt(hz))
          * @param imu_avg_ if we want to average the imu measurements (IJRR paper did not do this)
          */
-        CpiV1(double sigma_w, double sigma_wb, double sigma_a, double sigma_ab, bool imu_avg_ = false) :
+        CpiV1(float sigma_w, float sigma_wb, float sigma_a, float sigma_ab, bool imu_avg_ = false) :
                 CpiBase(sigma_w, sigma_wb, sigma_a, sigma_ab, imu_avg_) {}
 
 
@@ -80,13 +80,13 @@ namespace ov_core {
          * We will first analytically integrate our meansurements and Jacobians.
          * Then we perform numerical integration for our measurement covariance.
          */
-        void feed_IMU(double t_0, double t_1, Eigen::Matrix<double, 3, 1> w_m_0, Eigen::Matrix<double, 3, 1> a_m_0,
-                      Eigen::Matrix<double, 3, 1> w_m_1 = Eigen::Matrix<double, 3, 1>::Zero(),
-                      Eigen::Matrix<double, 3, 1> a_m_1 = Eigen::Matrix<double, 3, 1>::Zero()) {
+        void feed_IMU(float t_0, float t_1, Eigen::Matrix<float, 3, 1> w_m_0, Eigen::Matrix<float, 3, 1> a_m_0,
+                      Eigen::Matrix<float, 3, 1> w_m_1 = Eigen::Matrix<float, 3, 1>::Zero(),
+                      Eigen::Matrix<float, 3, 1> a_m_1 = Eigen::Matrix<float, 3, 1>::Zero()) {
 
 
             // Get time difference
-            double delta_t = t_1 - t_0;
+            float delta_t = t_1 - t_0;
             DT += delta_t;
 
             //If no time has passed do nothing
@@ -95,8 +95,8 @@ namespace ov_core {
             }
 
             // Get estimated imu readings
-            Eigen::Matrix<double, 3, 1> w_hat = w_m_0 - b_w_lin;
-            Eigen::Matrix<double, 3, 1> a_hat = a_m_0 - b_a_lin;
+            Eigen::Matrix<float, 3, 1> w_hat = w_m_0 - b_w_lin;
+            Eigen::Matrix<float, 3, 1> a_hat = a_m_0 - b_a_lin;
 
             // If averaging, average
             if (imu_avg) {
@@ -107,29 +107,29 @@ namespace ov_core {
             }
 
             // Get angle change w*dt
-            Eigen::Matrix<double, 3, 1> w_hatdt = w_hat * delta_t;
+            Eigen::Matrix<float, 3, 1> w_hatdt = w_hat * delta_t;
 
             // Get entries of w_hat
-            double w_1 = w_hat(0, 0);
-            double w_2 = w_hat(1, 0);
-            double w_3 = w_hat(2, 0);
+            float w_1 = w_hat(0, 0);
+            float w_2 = w_hat(1, 0);
+            float w_3 = w_hat(2, 0);
 
             // Get magnitude of w and wdt
-            double mag_w = w_hat.norm();
-            double w_dt = mag_w * delta_t;
+            float mag_w = w_hat.norm();
+            float w_dt = mag_w * delta_t;
 
             // Threshold to determine if equations will be unstable
             bool small_w = (mag_w < 0.008726646);
 
             // Get some of the variables used in the preintegration equations
-            double dt_2 = pow(delta_t, 2);
-            double cos_wt = cos(w_dt);
-            double sin_wt = sin(w_dt);
+            float dt_2 = pow(delta_t, 2);
+            float cos_wt = cos(w_dt);
+            float sin_wt = sin(w_dt);
 
-            Eigen::Matrix<double, 3, 3> w_x = skew_x(w_hat);
-            Eigen::Matrix<double, 3, 3> a_x = skew_x(a_hat);
-            Eigen::Matrix<double, 3, 3> w_tx = skew_x(w_hatdt);
-            Eigen::Matrix<double, 3, 3> w_x_2 = w_x * w_x;
+            Eigen::Matrix<float, 3, 3> w_x = skew_x(w_hat);
+            Eigen::Matrix<float, 3, 3> a_x = skew_x(a_hat);
+            Eigen::Matrix<float, 3, 3> w_tx = skew_x(w_hatdt);
+            Eigen::Matrix<float, 3, 3> w_x_2 = w_x * w_x;
 
 
             //==========================================================================
@@ -137,18 +137,18 @@ namespace ov_core {
             //==========================================================================
 
             // Get relative rotation
-            Eigen::Matrix<double, 3, 3> R_tau2tau1 = small_w ? eye3 - delta_t * w_x + (pow(delta_t, 2) / 2) * w_x_2 :
+            Eigen::Matrix<float, 3, 3> R_tau2tau1 = small_w ? eye3 - delta_t * w_x + (pow(delta_t, 2) / 2) * w_x_2 :
                                                      eye3 - (sin_wt / mag_w) * w_x + ((1.0 - cos_wt) / (pow(mag_w, 2.0))) * w_x_2;
 
             // Updated rotation and its transpose
-            Eigen::Matrix<double, 3, 3> R_k2tau1 = R_tau2tau1 * R_k2tau;
-            Eigen::Matrix<double, 3, 3> R_tau12k = R_k2tau1.transpose();
+            Eigen::Matrix<float, 3, 3> R_k2tau1 = R_tau2tau1 * R_k2tau;
+            Eigen::Matrix<float, 3, 3> R_tau12k = R_k2tau1.transpose();
 
             //Intermediate variables for evaluating the measurement/bias Jacobian update
-            double f_1;
-            double f_2;
-            double f_3;
-            double f_4;
+            float f_1;
+            float f_2;
+            float f_3;
+            float f_4;
 
             if (small_w) {
                 f_1 = -(pow(delta_t, 3) / 3);
@@ -163,12 +163,12 @@ namespace ov_core {
             }
 
             // Compute the main part of our analytical means
-            Eigen::Matrix<double, 3, 3> alpha_arg = ((dt_2 / 2.0) * eye3 + f_1 * w_x + f_2 * w_x_2);
-            Eigen::Matrix<double, 3, 3> Beta_arg = (delta_t * eye3 + f_3 * w_x + f_4 * w_x_2);
+            Eigen::Matrix<float, 3, 3> alpha_arg = ((dt_2 / 2.0) * eye3 + f_1 * w_x + f_2 * w_x_2);
+            Eigen::Matrix<float, 3, 3> Beta_arg = (delta_t * eye3 + f_3 * w_x + f_4 * w_x_2);
 
             // Matrices that will multiply the a_hat in the update expressions
-            Eigen::MatrixXd H_al = R_tau12k * alpha_arg;
-            Eigen::MatrixXd H_be = R_tau12k * Beta_arg;
+            Eigen::MatrixXf H_al = R_tau12k * alpha_arg;
+            Eigen::MatrixXf H_be = R_tau12k * Beta_arg;
 
             // Update the measurement means
             alpha_tau += beta_tau * delta_t + H_al * a_hat;
@@ -180,7 +180,7 @@ namespace ov_core {
             //==========================================================================
 
             // Get right Jacobian
-            Eigen::Matrix<double, 3, 3> J_r_tau1 = small_w ? eye3 - .5 * w_tx + (1.0 / 6.0) * w_tx * w_tx :
+            Eigen::Matrix<float, 3, 3> J_r_tau1 = small_w ? eye3 - .5 * w_tx + (1.0 / 6.0) * w_tx * w_tx :
                                                    eye3 - ((1 - cos_wt) / (pow((w_dt), 2.0))) * w_tx +
                                                    ((w_dt - sin_wt) / (pow(w_dt, 3.0))) * w_tx * w_tx;
 
@@ -193,65 +193,65 @@ namespace ov_core {
             H_b -= H_be;
 
             // Derivatives of R_tau12k wrt bias_w entries
-            Eigen::MatrixXd d_R_bw_1 = -R_tau12k * skew_x(J_q * e_1);
-            Eigen::MatrixXd d_R_bw_2 = -R_tau12k * skew_x(J_q * e_2);
-            Eigen::MatrixXd d_R_bw_3 = -R_tau12k * skew_x(J_q * e_3);
+            Eigen::MatrixXf d_R_bw_1 = -R_tau12k * skew_x(J_q * e_1);
+            Eigen::MatrixXf d_R_bw_2 = -R_tau12k * skew_x(J_q * e_2);
+            Eigen::MatrixXf d_R_bw_3 = -R_tau12k * skew_x(J_q * e_3);
 
             // Now compute the gyro bias Jacobian terms
-            double df_1_dbw_1;
-            double df_1_dbw_2;
-            double df_1_dbw_3;
+            float df_1_dbw_1;
+            float df_1_dbw_2;
+            float df_1_dbw_3;
 
-            double df_2_dbw_1;
-            double df_2_dbw_2;
-            double df_2_dbw_3;
+            float df_2_dbw_1;
+            float df_2_dbw_2;
+            float df_2_dbw_3;
 
-            double df_3_dbw_1;
-            double df_3_dbw_2;
-            double df_3_dbw_3;
+            float df_3_dbw_1;
+            float df_3_dbw_2;
+            float df_3_dbw_3;
 
-            double df_4_dbw_1;
-            double df_4_dbw_2;
-            double df_4_dbw_3;
+            float df_4_dbw_1;
+            float df_4_dbw_2;
+            float df_4_dbw_3;
 
             if (small_w) {
-                double df_1_dw_mag = -(pow(delta_t, 5) / 15);
+                float df_1_dw_mag = -(pow(delta_t, 5) / 15);
                 df_1_dbw_1 = w_1 * df_1_dw_mag;
                 df_1_dbw_2 = w_2 * df_1_dw_mag;
                 df_1_dbw_3 = w_3 * df_1_dw_mag;
 
-                double df_2_dw_mag = (pow(delta_t, 6) / 72);
+                float df_2_dw_mag = (pow(delta_t, 6) / 72);
                 df_2_dbw_1 = w_1 * df_2_dw_mag;
                 df_2_dbw_2 = w_2 * df_2_dw_mag;
                 df_2_dbw_3 = w_3 * df_2_dw_mag;
 
-                double df_3_dw_mag = -(pow(delta_t, 4) / 12);
+                float df_3_dw_mag = -(pow(delta_t, 4) / 12);
                 df_3_dbw_1 = w_1 * df_3_dw_mag;
                 df_3_dbw_2 = w_2 * df_3_dw_mag;
                 df_3_dbw_3 = w_3 * df_3_dw_mag;
 
-                double df_4_dw_mag = (pow(delta_t, 5) / 60);
+                float df_4_dw_mag = (pow(delta_t, 5) / 60);
                 df_4_dbw_1 = w_1 * df_4_dw_mag;
                 df_4_dbw_2 = w_2 * df_4_dw_mag;
                 df_4_dbw_3 = w_3 * df_4_dw_mag;
             } else {
-                double df_1_dw_mag = (pow(w_dt, 2) * sin_wt - 3 * sin_wt + 3 * w_dt * cos_wt) / pow(mag_w, 5);
+                float df_1_dw_mag = (pow(w_dt, 2) * sin_wt - 3 * sin_wt + 3 * w_dt * cos_wt) / pow(mag_w, 5);
                 df_1_dbw_1 = w_1 * df_1_dw_mag;
                 df_1_dbw_2 = w_2 * df_1_dw_mag;
                 df_1_dbw_3 = w_3 * df_1_dw_mag;
 
 
-                double df_2_dw_mag = (pow(w_dt, 2) - 4 * cos_wt - 4 * w_dt * sin_wt + pow(w_dt, 2) * cos_wt + 4) / (pow(mag_w, 6));
+                float df_2_dw_mag = (pow(w_dt, 2) - 4 * cos_wt - 4 * w_dt * sin_wt + pow(w_dt, 2) * cos_wt + 4) / (pow(mag_w, 6));
                 df_2_dbw_1 = w_1 * df_2_dw_mag;
                 df_2_dbw_2 = w_2 * df_2_dw_mag;
                 df_2_dbw_3 = w_3 * df_2_dw_mag;
 
-                double df_3_dw_mag = (2 * (cos_wt - 1) + w_dt * sin_wt) / (pow(mag_w, 4));
+                float df_3_dw_mag = (2 * (cos_wt - 1) + w_dt * sin_wt) / (pow(mag_w, 4));
                 df_3_dbw_1 = w_1 * df_3_dw_mag;
                 df_3_dbw_2 = w_2 * df_3_dw_mag;
                 df_3_dbw_3 = w_3 * df_3_dw_mag;
 
-                double df_4_dw_mag = (2 * w_dt + w_dt * cos_wt - 3 * sin_wt) / (pow(mag_w, 5));
+                float df_4_dw_mag = (2 * w_dt + w_dt * cos_wt - 3 * sin_wt) / (pow(mag_w, 5));
                 df_4_dbw_1 = w_1 * df_4_dw_mag;
                 df_4_dbw_2 = w_2 * df_4_dw_mag;
                 df_4_dbw_3 = w_3 * df_4_dw_mag;
@@ -284,7 +284,7 @@ namespace ov_core {
             //==========================================================================
 
             // Going to need orientation at intermediate time i.e. at .5*dt;
-            Eigen::Matrix<double, 3, 3> R_mid = small_w ? eye3 - .5 * delta_t * w_x + (pow(.5 * delta_t, 2) / 2) * w_x_2 :
+            Eigen::Matrix<float, 3, 3> R_mid = small_w ? eye3 - .5 * delta_t * w_x + (pow(.5 * delta_t, 2) / 2) * w_x_2 :
                                                 eye3 - (sin(mag_w * .5 * delta_t) / mag_w) * w_x + ((1.0 - cos(mag_w * .5 * delta_t)) / (pow(mag_w, 2.0))) * w_x_2;
             R_mid = R_mid * R_k2tau;
 
@@ -293,7 +293,7 @@ namespace ov_core {
             //k1-------------------------------------------------------------------------------------------------
 
             // Build state Jacobian
-            Eigen::Matrix<double, 15, 15> F_k1 = Eigen::Matrix<double, 15, 15>::Zero();
+            Eigen::Matrix<float, 15, 15> F_k1 = Eigen::Matrix<float, 15, 15>::Zero();
             F_k1.block(0, 0, 3, 3) = -w_x;
             F_k1.block(0, 3, 3, 3) = -eye3;
             F_k1.block(6, 0, 3, 3) = -R_k2tau.transpose() * a_x;
@@ -301,20 +301,20 @@ namespace ov_core {
             F_k1.block(12, 6, 3, 3) = eye3;
 
             // Build noise Jacobian
-            Eigen::Matrix<double, 15, 12> G_k1 = Eigen::Matrix<double, 15, 12>::Zero();
+            Eigen::Matrix<float, 15, 12> G_k1 = Eigen::Matrix<float, 15, 12>::Zero();
             G_k1.block(0, 0, 3, 3) = -eye3;
             G_k1.block(3, 3, 3, 3) = eye3;
             G_k1.block(6, 6, 3, 3) = -R_k2tau.transpose();
             G_k1.block(9, 9, 3, 3) = eye3;
 
             // Get covariance derivative
-            Eigen::Matrix<double, 15, 15> P_dot_k1 = F_k1 * P_meas + P_meas * F_k1.transpose() + G_k1 * Q_c * G_k1.transpose();
+            Eigen::Matrix<float, 15, 15> P_dot_k1 = F_k1 * P_meas + P_meas * F_k1.transpose() + G_k1 * Q_c * G_k1.transpose();
 
 
             //k2-------------------------------------------------------------------------------------------------
 
             // Build state Jacobian
-            Eigen::Matrix<double, 15, 15> F_k2 = Eigen::Matrix<double, 15, 15>::Zero();
+            Eigen::Matrix<float, 15, 15> F_k2 = Eigen::Matrix<float, 15, 15>::Zero();
             F_k2.block(0, 0, 3, 3) = -w_x;
             F_k2.block(0, 3, 3, 3) = -eye3;
             F_k2.block(6, 0, 3, 3) = -R_mid.transpose() * a_x;
@@ -322,31 +322,31 @@ namespace ov_core {
             F_k2.block(12, 6, 3, 3) = eye3;
 
             // Build noise Jacobian
-            Eigen::Matrix<double, 15, 12> G_k2 = Eigen::Matrix<double, 15, 12>::Zero();
+            Eigen::Matrix<float, 15, 12> G_k2 = Eigen::Matrix<float, 15, 12>::Zero();
             G_k2.block(0, 0, 3, 3) = -eye3;
             G_k2.block(3, 3, 3, 3) = eye3;
             G_k2.block(6, 6, 3, 3) = -R_mid.transpose();
             G_k2.block(9, 9, 3, 3) = eye3;
 
             // Get covariance derivative
-            Eigen::Matrix<double, 15, 15> P_k2 = P_meas + P_dot_k1 * delta_t / 2.0;
-            Eigen::Matrix<double, 15, 15> P_dot_k2 = F_k2 * P_k2 + P_k2 * F_k2.transpose() + G_k2 * Q_c * G_k2.transpose();
+            Eigen::Matrix<float, 15, 15> P_k2 = P_meas + P_dot_k1 * delta_t / 2.0;
+            Eigen::Matrix<float, 15, 15> P_dot_k2 = F_k2 * P_k2 + P_k2 * F_k2.transpose() + G_k2 * Q_c * G_k2.transpose();
 
             //k3-------------------------------------------------------------------------------------------------
 
             // Our state and noise Jacobians are the same as k2
             // Since k2 and k3 correspond to the same estimates for the midpoint
-            Eigen::Matrix<double, 15, 15> F_k3 = F_k2;
-            Eigen::Matrix<double, 15, 12> G_k3 = G_k2;
+            Eigen::Matrix<float, 15, 15> F_k3 = F_k2;
+            Eigen::Matrix<float, 15, 12> G_k3 = G_k2;
 
             // Get covariance derivative
-            Eigen::Matrix<double, 15, 15> P_k3 = P_meas + P_dot_k2 * delta_t / 2.0;
-            Eigen::Matrix<double, 15, 15> P_dot_k3 = F_k3 * P_k3 + P_k3 * F_k3.transpose() + G_k3 * Q_c * G_k3.transpose();
+            Eigen::Matrix<float, 15, 15> P_k3 = P_meas + P_dot_k2 * delta_t / 2.0;
+            Eigen::Matrix<float, 15, 15> P_dot_k3 = F_k3 * P_k3 + P_k3 * F_k3.transpose() + G_k3 * Q_c * G_k3.transpose();
 
             //k4-------------------------------------------------------------------------------------------------
 
             // Build state Jacobian
-            Eigen::Matrix<double, 15, 15> F_k4 = Eigen::Matrix<double, 15, 15>::Zero();
+            Eigen::Matrix<float, 15, 15> F_k4 = Eigen::Matrix<float, 15, 15>::Zero();
             F_k4.block(0, 0, 3, 3) = -w_x;
             F_k4.block(0, 3, 3, 3) = -eye3;
             F_k4.block(6, 0, 3, 3) = -R_k2tau1.transpose() * a_x;
@@ -354,15 +354,15 @@ namespace ov_core {
             F_k4.block(12, 6, 3, 3) = eye3;
 
             // Build noise Jacobian
-            Eigen::Matrix<double, 15, 12> G_k4 = Eigen::Matrix<double, 15, 12>::Zero();
+            Eigen::Matrix<float, 15, 12> G_k4 = Eigen::Matrix<float, 15, 12>::Zero();
             G_k4.block(0, 0, 3, 3) = -eye3;
             G_k4.block(3, 3, 3, 3) = eye3;
             G_k4.block(6, 6, 3, 3) = -R_k2tau1.transpose();
             G_k4.block(9, 9, 3, 3) = eye3;
 
             // Get covariance derivative
-            Eigen::Matrix<double, 15, 15> P_k4 = P_meas + P_dot_k3 * delta_t;
-            Eigen::Matrix<double, 15, 15> P_dot_k4 = F_k4 * P_k4 + P_k4 * F_k4.transpose() + G_k4 * Q_c * G_k4.transpose();
+            Eigen::Matrix<float, 15, 15> P_k4 = P_meas + P_dot_k3 * delta_t;
+            Eigen::Matrix<float, 15, 15> P_dot_k4 = F_k4 * P_k4 + P_k4 * F_k4.transpose() + G_k4 * Q_c * G_k4.transpose();
 
 
             //done-------------------------------------------------------------------------------------------------
