+++ ov_core/src/track/TrackSIM.cpp
@@ -26,7 +26,7 @@ using namespace ov_core;
 
 
 
-void TrackSIM::feed_measurement_simulation(double timestamp, const std::vector<int> &camids, const std::vector<std::vector<std::pair<size_t,Eigen::VectorXf>>> &feats) {
+void TrackSIM::feed_measurement_simulation(float timestamp, const std::vector<int> &camids, const std::vector<std::vector<std::pair<size_t,Eigen::VectorXf>>> &feats) {
 
 
     // Assert our two vectors are equal
